backend/scraper/requirements-test.txt
playwright>=1.40.0
requests>=2.31.0
coverage>=7.0.0
python-dotenv>=1.0.0